requirements.txt
Django~=3.2.6
django-ckeditor>=1.2.2